chore: extract websocket server into separate package
