+++ apps/server/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@ffmpeg-installer/ffmpeg": "^1.1.0",
     "@ffprobe-installer/ffprobe": "^2.1.2",
+    "@react-native-templates/ws-server": "0.1.0",
     "better-sqlite3": "^9.4.0",
     "body-parser": "^1.20.2",
     "chroma-js": "^2.4.2",
