+++ apps/server/tsconfig.json
@@ -14,5 +14,7 @@
     "path": "../../packages/core"
   }, {
     "path": "../../packages/ws-client"
+  }, {
+    "path": "../../packages/ws-server"
   }]
 }
