+++ packages/ws-server/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@react-native-templates/ws-server",
+  "version": "0.1.0",
+  "license": "Apache-2.0",
+  "description": "",
+  "main": "lib",
+  "react-native": "src",
+  "files": [
+    "src",
+    "lib",
+    "!**/__tests__",
+    "!**/__fixtures__",
+    "!**/__mocks__",
+    "!**/.*"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ericvicenti/react-native-templates.git",
+    "directory": "packages/ws-server"
+  },
+  "author": "Mike Grabowski <grabbou@gmail.com> (https://github.com/grabbou)",
+  "bugs": {
+    "url": "https://github.com/ericvicenti/react-native-templates/issues"
+  },
+  "homepage": "https://github.com/ericvicenti/react-native-templates#readme",
+  "dependencies": {
+    "ws": "^8.16.0"
+  },
+  "devDependencies": {
+    "@react-native-templates/core": "0.1.0",
+    "@react-native-templates/ws-client": "0.1.0",
+    "metro-react-native-babel-preset": "^0.77.0"
+  }
+}
