+++ packages/ws-server/src/index.ts
@@ -1,0 +1,1 @@
+export * from './ws-rnt-server'
