+++ packages/ws-server/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./lib"
+  },
+  "references": [
+    {
+      "path": "../core"
+    },
+    {
+      "path": "../ws-client"
+    }
+  ]
+}
